% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quartopost.R
\name{quartopost}
\alias{quartopost}
\title{Create a new quarto blog post}
\usage{
quartopost()
}
\value{
Nothing. What matters are the side effects:
\itemize{
\item create a folder in kebab notation with the name of date followed by the post title
\item create a "index.qmd" file
\item copy the (optional) image file into the folder
\item populate the YAML field
\item open the file for editing
}
}
\description{
\code{quartopost()} opens a dialog window to input title and other
desired data to create the file for the blog post.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
quartopost()
\dontshow{\}) # examplesIf}
}
